backend/requirements.txt
Pillow # image
validate_email # valid email
pyDNS # http://pydns.sourceforge.net/
# psycopg2 # Postgres, ubuntu use pip install, mac download app(http://postgresapp.com) 
gunicorn # https://github.com/benoitc/gunicorn(a WSGI HTTP Server for UNIX)
djangorestframework # django rest framework
pygments  # We'll be using this for the code highlighting
markdown # https://github.com/klen/django_markdown(rest framework)
django-filter # https://github.com/carltongibson/django-filter(rest framework)
httpie # https://github.com/jkbrzt/httpie#installation (rest framework)
coreapi # https://github.com/core-api/python-client
coreapi-cli # coreapi get http://127.0.0.1:8000/schema/; coreapi action users list
